+++ mapentity/static/mapentity/mapentity.filter.js
@@ -9,6 +9,7 @@ MapEntity.TogglableFilter = L.Class.extend({
 
         this.fields = {};
         this.visible = false;
+        this.loaded_form = false;
         this.popover = $('#filters-popover')
                           .popover({
                               placement: 'right',
@@ -29,26 +30,6 @@ MapEntity.TogglableFilter = L.Class.extend({
 
         $('#mainfilter').find('select,input').change(function (e) {
             self.setfield(this);
-        });
-
-        // Use chosen for multiple values
-        $("form#mainfilter").bind("reset", function() {
-            setTimeout(function() {
-                $('form#mainfilter select[multiple]').trigger('chosen:updated');
-            }, 1);
-        });
-
-        // Make sure filter-set class is added if a choice is selected.
-        $('#mainfilter select[multiple]').chosen().on('change', function (e) {
-            var $target = $(e.target),
-                name = $target.attr('name'),
-                $container = $('div#id_' + name + '_chzn > ul');
-            if ($(e.target).find('option:selected').length > 0) {
-                $container.addClass('filter-set');
-            }
-            else {
-                $container.removeClass('filter-set');
-            }
         });
 
 
@@ -85,6 +66,58 @@ MapEntity.TogglableFilter = L.Class.extend({
 
     tip: function () {
         return $(this.popover.data('bs.popover').tip);
+    },
+
+    load_filter_form: function (mapsync) {
+        var self = this;
+        // On first click on Filter button, load HTML content for form
+        var $mainfilter = $('#mainfilter');
+        if (!self.loaded_form) {
+            var filter_url = $mainfilter.attr('filter-url');
+            $.get(filter_url)
+                .done(response => {
+                    $('#filters-panel .filter-spinner-container').hide();
+                    // Replace simple form that contains only BBOX with full form, including attributes
+                    $mainfilter.replaceWith(response);
+                    $mainfilter = $('#mainfilter');
+                    // Update L.MapListSync to refresh datatable on change
+                    mapsync.options.filter.form = $mainfilter;
+                    // Bind new form buttons to keep refreshing list on click
+                    $("#filter").click(function(e) {
+                        mapsync._onFormSubmit(e);
+                    });
+                    $("#reset").click(function(e) {
+                        mapsync._onFormReset(e);
+                    });
+                    // Bind setfields to update list of enabled fields displayed on hover
+                    $mainfilter.find('select,input').change(function () {
+                        self.setfield(this);
+                    });
+                    self.loaded_form = true;
+
+                    // Use chosen for multiple values
+                    $mainfilter.bind("reset", function() {
+                        setTimeout(function() {
+                            $mainfilter.find('select[multiple]').trigger('chosen:updated');
+                        }, 1);
+                    });
+
+                    // Make sure filter-set class is added if a choice is selected.
+                    $mainfilter.find('select[multiple]').chosen().on('change', function (e) {
+                        var $target = $(e.target),
+                            name = $target.attr('name'),
+                            $container = $('div#id_' + name + '_chzn > ul');
+                        var hasSelectedOption = $target.find('option:selected').length > 0;
+                        $container.toggleClass('filter-set', hasSelectedOption);
+                    });
+                    // Move right-filters to right side
+                    $mainfilter.find('.right-filter').parent('p').detach().appendTo('#mainfilter > .right');
+                    // Trigger event allowing to launch further processing
+                    var context = $('body').data();
+                    $(window).trigger('entity:view:filter', {modelname: context.modelname});
+                })
+                .fail(xhr => console.error('Error:', xhr.status));
+        }
     },
 
     showinfo: function () {